requirements.txt
Django==5.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-gis==1.1
djangorestframework_simplejwt==5.4.0
httpie==3.2.4
geojson==3.2.0